fix: enhance wallet binding flow
